qcom-git-stats: use yaml metadata to define identities
Define a yaml structure to describe developers metadata / identities,
which contains:
* developer's name
* one or more emails
* one or more 'enrollments' or groups

Groups will be used later to filter in/out developers when analysing
the commits. Each enrollment has a 'start' date, a commit will belong
to a group if the commit date is after the start date of a group.

In order to simplify arguments, argparse is now used. Identities can
be split into multiple files.

Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>