git-stats: improve support for organization
When the identity file defines an 'organization', we end up with
matching email addresses but without any corresponding 'name' in the
identity file. In this situation let's default to using the name from
the commit itself.

For signed-off-by data, we need to catch the name on top of the
email. If the developer who signed-off is part of one of the
organization tracked in the identity file (e.g. if the email address
matches) then we will need to use the name found in the signed-off-by
tag.

Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>