git-stats: add support for group 'end' date
In the configuration file, developers may be affiliated to a group (or
team/organization). Up to now, we only rely on a 'start' date to
detect if a developer belongs to a group. Add an optional 'end' date
as well. If the end data is set, the commit will be accounted for only
if the developer was in the given during during that period.

Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>