git-stats: update the logic when looking for the subsystems
We were starting straight from the 'parent' of the changed file,
however if we want to catch specific files (such as MAINTAINERS), we
should not strip the file name in the first place. So let's change the
logic so that we can check the filename before trying to look for a
Makefile.

Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>