git-stats: improve regex support
In edacea4ce88277d10e4f4cf993488aceda6f47a4, we fixed support for
email addresses with special characters (such as john+doe@xx). But it
broke support for reg exp which we had already. e.g. it was working
so far to have .*@linaro.org in the YAML config file, which no longer
works.

In order to support both emails with special chars and regex, we need
to clearly identify regex in the config file. So let's use /<>/
markers for that.

Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>